tests/wpt/web-platform-tests/XMLHttpRequest/preserve-ua-header-on-redirect.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: User-Agent header is preserved on redirect</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      var ua = "Servo/wptrunner";
      var test = async_test()
      test.step(function() {
        var client = new XMLHttpRequest()
          client.onreadystatechange = function() {
            test.step(function() {
              if(client.readyState == 4) {
                assert_equals(client.responseText, "user-agent,")
                test.done()
              }
            })
          }
        client.open("POST", "resources/redirect.py?location="+encodeURIComponent("inspect-headers.py?filter_value="+ua))
        client.send(null)
      })
    </script>
  </body>
</html>
